fix a problem with breaks due to an unreachable statement and fix some warnings with the Intel compiler

+++ src/lib/WP1LeftIndentGroup.cpp
@@ -46,5 +46,5 @@ void WP1LeftIndentGroup::_readContents(WPXInputStream *input)
 void WP1LeftIndentGroup::parse(WP1Listener *listener)
 {
 	WPD_DEBUG_MSG(("WordPerfect: handling the Left Indent group\n"));
-	listener->leftIndent(m_newLeftMargin - m_oldLeftMargin);
+	listener->leftIndent((uint16_t)(m_newLeftMargin - m_oldLeftMargin));
 }
